requirements.txt
contourpy==1.1.0
cycler==0.11.0
dt-apriltags==3.1.7
fonttools==4.42.1
importlib-resources==6.0.1
kiwisolver==1.4.5
matplotlib==3.7.3
numpy==1.24.4
opencv-python==4.8.0.76
packaging==23.1
Pillow==10.0.1
pyparsing==3.1.1
python-dateutil==2.8.2
six==1.16.0
zipp==3.16.2
